chore(tests): analyzer tests in river dir, - mocks
