+++ tests/river/unit/analyzer/test_attribute.py
