+++ tests/river/unit/analyzer/test_condition.py
