+++ tests/river/unit/analyzer/test_input_groups.py
