+++ tests/river/unit/analyzer/test_merging_script.py
