+++ tests/river/unit/analyzer/test_sql_column.py
